Implement reprotect method on virtual memory manager (currently stubbed)
