--- src/Ryujinx.HLE/HOS/Kernel/Memory/MemoryPermission.cs
@@ -1,20 +1,0 @@
-using System;
-
-namespace Ryujinx.HLE.HOS.Kernel.Memory
-{
-    [Flags]
-    enum KMemoryPermission : uint
-    {
-        None = 0,
-        UserMask = Read | Write | Execute,
-        Mask = uint.MaxValue,
-
-        Read = 1 << 0,
-        Write = 1 << 1,
-        Execute = 1 << 2,
-        DontCare = 1 << 28,
-
-        ReadAndWrite = Read | Write,
-        ReadAndExecute = Read | Execute,
-    }
-}
